Update README.md
Any firm that wants to succeed in the fast-paced business world of today needs to have an effective supply chain and logistics system. Companies are increasingly turning to Enterprise Resource Planning (ERP) technologies to meet this requirement. Odoo, an open-source ERP software, has emerged as a major changer in the logistics and supply chain management sectors. The current version, Odoo 17 ERP, is modernizing and improving important business functions.